smsudata-dronepol-pipeline/requirements.txt
pandas
mysql-connector-python
psycopg2-binary
python-dotenv